04-analytics-engineering/requirements.txt
pandas
google-cloud-storage
requests
pyarrow